+++ README.md
@@ -2,11 +2,10 @@
 
 [![NPM version][npm-image]][npm-url]
 [![NPM downloads][downloads-image]][downloads-url]
-![badge-size-es5-url]
-![badge-size-es2015-url]
 [![Build status][travis-image]][travis-url]
 [![Coverage Status][coveralls-image]][coveralls-url]
 [![Greenkeeper badge](https://badges.greenkeeper.io/unional/color-map.svg)](https://greenkeeper.io/)
+![badge-size-es5-url]
 
 Color map generator with no dependency.
 
@@ -106,4 +105,3 @@ Generated by [`generator-unional@0.3.1`](https://github.com/unional/unional-cli)
 [coveralls-image]: https://coveralls.io/repos/github/unional/color-map/badge.svg
 [coveralls-url]: https://coveralls.io/github/unional/color-map
 [badge-size-es5-url]: http://img.badgesize.io/unional/color-map/master/dist/color-map.es5.js.svg?label=es5_size
-[badge-size-es2015-url]: http://img.badgesize.io/unional/color-map/master/dist/color-map.es2015.js.svg?label=es2015_size
