+++ tsconfig.json
@@ -6,6 +6,6 @@
   "buildOnSave": false,
   "compileOnSave": false,
   "include": [
-    "src/**/*"
+    "src"
   ]
 }
